backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.0.0
cryptography==41.0.4
gunicorn==21.2.0
pandas==2.0.3
openpyxl==3.1.2

# Development dependencies
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0